chore: split `PluginScreenPlatform` and `AboutScreenPlatform`
