+++ composeApp/src/androidMain/kotlin/fr/husi/ui/PluginScreenPlatform.android.kt
@@ -1,17 +1,13 @@
 package fr.husi.ui
 
-import android.annotation.SuppressLint
-import android.content.Context
 import android.content.Intent
 import android.content.pm.PackageInfo
 import android.net.Uri
 import android.os.Build
-import android.os.PowerManager
 import android.provider.Settings
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.platform.LocalContext
-import androidx.core.net.toUri
 import fr.husi.fmt.PluginEntry
 import fr.husi.ktx.Logs
 import fr.husi.plugin.Plugins
@@ -70,30 +66,5 @@ private fun PackageInfo.versionCodeCompat(): Long {
     } else {
         @Suppress("DEPRECATION")
         versionCode.toLong()
-    }
-}
-
-@Composable
-internal actual fun rememberShouldRequestBatteryOptimizations(): Boolean {
-    val context = LocalContext.current
-    val powerManger = context.getSystemService(Context.POWER_SERVICE) as PowerManager
-    return remember(context, powerManger) {
-        !powerManger.isIgnoringBatteryOptimizations(context.packageName)
-    }
-}
-
-@SuppressLint("BatteryLife")
-@Composable
-internal actual fun rememberRequestIgnoreBatteryOptimizations(): () -> Unit {
-    val context = LocalContext.current
-    return remember(context) {
-        {
-            context.startActivity(
-                Intent()
-                    .setAction(Settings.ACTION_REQUEST_IGNORE_BATTERY_OPTIMIZATIONS)
-                    .setData("package:${context.packageName}".toUri())
-                    .addFlags(Intent.FLAG_ACTIVITY_NEW_TASK),
-            )
-        }
     }
 }
