+++ composeApp/src/commonMain/kotlin/fr/husi/ui/AboutScreenPlatform.kt
@@ -1,0 +1,9 @@
+package fr.husi.ui
+
+import androidx.compose.runtime.Composable
+
+@Composable
+internal expect fun rememberShouldRequestBatteryOptimizations(): Boolean
+
+@Composable
+internal expect fun rememberRequestIgnoreBatteryOptimizations(): () -> Unit
