+++ composeApp/src/commonMain/kotlin/fr/husi/ui/PluginScreenPlatform.kt
@@ -7,9 +7,3 @@ internal expect fun platformPluginsFlow(): Flow<List<PluginDisplay>>
 
 @Composable
 internal expect fun rememberOpenPluginCard(): (PluginDisplay) -> Unit
-
-@Composable
-internal expect fun rememberShouldRequestBatteryOptimizations(): Boolean
-
-@Composable
-internal expect fun rememberRequestIgnoreBatteryOptimizations(): () -> Unit
